feat: move markdown handling to common
