+++ src/psycop/common/model_evaluation/markdown/test_md_objects.py
@@ -3,7 +3,7 @@ from pathlib import Path
 import pandas as pd
 import pytest
 from psycop.common.test_utils.str_to_df import str_to_df
-from psycop.projects.t2d.paper_outputs.aggregate_eval.md_objects import (
+from psycop.common.model_evaluation.markdown.md_objects import (
     MarkdownFigure,
     MarkdownTable,
     create_supplementary_from_markdown_artifacts,
